+++ src/Minimization/Heuristic/Heuristic.jl
@@ -34,7 +34,7 @@ using MatrixBandwidth: connected_components
 using MatrixBandwidth: _requires_structural_symmetry
 
 using MatrixBandwidth.Minimization
-using MatrixBandwidth.Minimization: _approach, _bool_minimal_band_ordering
+using MatrixBandwidth.Minimization: _approach, _minimize_bandwidth_impl
 
 using DataStructures: Queue