+++ src/Minimization/Metaheuristic/Metaheuristic.jl
@@ -32,7 +32,7 @@ using MatrixBandwidth: NotImplementedError
 using MatrixBandwidth: _requires_structural_symmetry
 
 using MatrixBandwidth.Minimization
-using MatrixBandwidth.Minimization: _approach, _bool_minimal_band_ordering
+using MatrixBandwidth.Minimization: _approach, _minimize_bandwidth_impl
 
 export
     # Types